backend/requirements.txt
flask==3.1.1
flask_sqlalchemy==3.1.1
flask_restful==0.3.10
flask-jwt-extended==4.7.1
python-dotenv==1.1.1
python-dateutil==2.9.0.post0